COM-2682 - send history data in batch + move constants
