--- frontend/helpers/constants.tsx
@@ -1,5 +1,0 @@
-export const RENT = 'rent';
-export const RENTAL_EXPENSES = 'rentalExpenses';
-export const CURRENT_EXPENSES = 'currentExpenses';
-
-export const AMOUNTS_NATURE = [RENT, RENTAL_EXPENSES, CURRENT_EXPENSES];
