+++ frontend/helpers/sepa.tsx
@@ -8,8 +8,18 @@ import {
   getRoommatesData,
 } from './airtable';
 import { addMessageAndThrow } from './errors';
-import { AMOUNTS_NATURE, CURRENT_EXPENSES, RENT, RENTAL_EXPENSES } from './constants';
-import { HISTORY_AMOUNT_FIELD_ID, HISTORY_DATE_FIELD_ID, HISTORY_DEBITOR_NAME_FIELD_ID, HISTORY_IBAN_FIELD_ID, HISTORY_RUM_FIELD_ID, HISTORY_TABLE, HISTORY_TRANSACTION_ID_FIELD_ID, HISTORY_TRANSACTION_NUMBER_FIELD_ID, HISTORY_TYPE_FIELD_ID } from '../../.env/models';
+import { AMOUNTS_NATURE, CURRENT_EXPENSES, RENT, RENTAL_EXPENSES } from '../data/constants';
+import {
+  HISTORY_AMOUNT_FIELD_ID,
+  HISTORY_DATE_FIELD_ID,
+  HISTORY_DEBITOR_NAME_FIELD_ID,
+  HISTORY_IBAN_FIELD_ID,
+  HISTORY_RUM_FIELD_ID,
+  HISTORY_TABLE,
+  HISTORY_TRANSACTION_ID_FIELD_ID,
+  HISTORY_TRANSACTION_NUMBER_FIELD_ID,
+  HISTORY_TYPE_FIELD_ID,
+} from '../../.env/models';
 
 
 export const createXMLDocument = () => ({
